Finished Question 15. Quite happy with the solution.

+++ Project Euler/Question_14.cpp
@@ -2,14 +2,14 @@ using namespace std;
 #include <vector>
 #include "Question_14.h"
 
-long long Collatz (long long target, vector<int>& seen_array, int MAX_NUMBER) {
+long long Collatz (long long target, vector<long long>& seen_array, int MAX_NUMBER) {
 
 	long long new_target = target;
 
 	if (target == 1) {
 		return 0;
-	} else if (target < MAX_NUMBER && seen_array[target] != 0) {
-		return seen_array[target];
+	} else if (target < MAX_NUMBER && seen_array[(int) target] != 0) {
+		return seen_array[(int)target];
 	} else if (target % 2) {
 		new_target = ((3 * target) + 1);
 	} else {
@@ -17,22 +17,22 @@ long long Collatz (long long target, vector<int>& seen_array, int MAX_NUMBER) {
 	}
 
 	if (target < MAX_NUMBER) {
-		seen_array[target] = (1 + Collatz (new_target, seen_array, MAX_NUMBER));
-		return seen_array[target];
+		seen_array[(int) target] = (1 + Collatz (new_target, seen_array, MAX_NUMBER));
+		return seen_array[(int) target];
 	}
 	return (1 + Collatz (new_target, seen_array, MAX_NUMBER));
 }
 
-int Question_14 () {
+long long Question_14 () {
 
 	int MAX_NUMBER = 1000000;
 
-	vector <int> previously_seen (MAX_NUMBER);
-	int gsf = 1;
+	vector <long long> previously_seen ((int) MAX_NUMBER);
+	long long gsf = 1;
 	int gsf_i = 1;
 
 	for (int i = 1; i < MAX_NUMBER; i++) {
-		int next_seq = Collatz (i, previously_seen, MAX_NUMBER);
+		long long next_seq = Collatz (i, previously_seen, MAX_NUMBER);
 
 		if (next_seq > gsf) {
 			gsf = next_seq;
