+++ Project Euler/Question_14.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Finds the longest Collatz chain under 1,000,00.
-int Question_14 ();
+long long Question_14 ();
